requirements.txt
aiogram==2.23.1
flask
requests~=2.32.3
apscheduler
python-dotenv~=1.0.1
python-telegram-bot[job-queue]~=21.10
